move registry_files tests to its own modle
