+++ test_empty_registry.csv
@@ -1,0 +1,1 @@
+source,last_update,hash,error_message,output_path
